% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeExploreGGRaptR.R
\name{executeExploreGGRaptR}
\alias{executeExploreGGRaptR}
\title{Perform the required operations for displaying interactive plot generator.}
\usage{
executeExploreGGRaptR(df_name, df)
}
\description{
Time-stamp: <2017-08-10 17:07:41 Graham Williams>
}
