% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore.R
\name{executeExplorePlot}
\alias{executeExplorePlot}
\title{dens <- density(ds[ds$grp=="All",1], na.rm=TRUE)\n',
rs<- max(summary(as.factor(round(ds[ds$grp=="All",',
1], digits=2))))/max(dens$y)\n',
lines(dens$x, dens$y*rs, type="l")',}
\usage{
executeExplorePlot(dataset, boxplots = getSelectedVariables("boxplot"),
  hisplots = getSelectedVariables("hisplot"),
  cumplots = getSelectedVariables("cumplot"),
  benplots = getSelectedVariables("benplot"),
  barplots = getSelectedVariables("barplot"),
  dotplots = getSelectedVariables("dotplot"),
  mosplots = getSelectedVariables("mosplot"),
  paiplots = getSelectedVariables("paiplot"), stratify = TRUE,
  sampling = NULL, target = crs$target, newplot = TRUE)
}
\description{
dens <- density(ds[ds$grp=="All",1], na.rm=TRUE)\n',
rs<- max(summary(as.factor(round(ds[ds$grp=="All",',
1], digits=2))))/max(dens$y)\n',
lines(dens$x, dens$y*rs, type="l")',
}
