% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeModelRF.R
\name{executeModelRF}
\alias{executeModelRF}
\title{Build a random forest model: traditional or conditional.}
\usage{
executeModelRF(traditional = TRUE, conditional = !traditional)
}
\description{
Time-stamp: <2016-07-22 22:06:58 Graham Williams>
}
