% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/executeModelRxDTree.R
\name{executeModelRxDTree}
\alias{executeModelRxDTree}
\title{Build a decision tree based from xdf dataset.}
\usage{
executeModelRxDTree()
}
\description{
Time-stamp: <2017-08-18 12:14:38 Graham Williams>
}
