% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLibs.R
\name{loadLibs}
\alias{loadLibs}
\title{Load a list of libraries, reporting to the Rattle Log}
\usage{
loadLibs(l)
}
\arguments{
\item{l}{Vector of pairs, "package name" "used function".}
}
\value{
returns list of packages that get loaded.
}
\description{
Only load the package if not already loaded. If already loaded then
we don't return the name from the function.
}
