% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unloadLibs.R
\name{unloadLibs}
\alias{unloadLibs}
\title{Unload pacakges}
\usage{
unloadLibs(l)
}
\arguments{
\item{l}{Vector of package names.}
}
\value{
nothing.
}
\description{
Detach the list of pacakges, only detaching those that are on the
search path.
}
